+++ src/prelude.rs
@@ -1,11 +1,11 @@
 //! Commonly used items.
+pub use crate::error::{MjDataError, MjSceneError, MjEditError, MjModelError, MjVfsError};
 pub use crate::wrappers::mj_editing::{MjSpec, SpecItem};
 pub use crate::wrappers::mj_visualization::*;
 pub use crate::wrappers::mj_rendering::*;
 pub use crate::wrappers::mj_option::*;
 pub use crate::wrappers::mj_model::*;
 pub use crate::wrappers::mj_data::*;
-pub use crate::error::{MjDataError, MjSceneError, MjEditError, MjModelError, MjVfsError};
 #[cfg(any(feature = "viewer", feature = "renderer-winit-fallback"))]
 pub use crate::error::GlInitError;